Tests updated
